Remove v7 compat classes
Additionally, removed some methods that are not required to be in the compat classes, including:
- Methods that are now available on all supported platforms after the recent minsdkversion bump
- Methods that are supported on all platform through the support library and don't gain any benefits from being in our compat classes
