+++ AnkiDroid/src/main/java/com/ichi2/anki/UIUtils.java
@@ -4,10 +4,16 @@ package com.ichi2.anki;
 import android.app.Activity;
 import android.content.Context;
 
+import android.support.v7.app.ActionBar;
+import android.support.v7.app.ActionBarActivity;
+import android.text.SpannableStringBuilder;
+import android.text.style.CharacterStyle;
+import android.text.style.ForegroundColorSpan;
 import android.view.WindowManager;
 
 import com.ichi2.async.DeckTask;
 import com.ichi2.async.DeckTask.TaskData;
+import com.ichi2.themes.Themes;
 
 import java.util.Calendar;
 
@@ -67,4 +73,38 @@ public class UIUtils {
                 WindowManager.LayoutParams.FLAG_FULLSCREEN);
     }
 
+
+    public static void setTitle(Activity activity, String title) {
+        ActionBarActivity actionBarActivity = (ActionBarActivity) activity;
+        ActionBar actionBar = actionBarActivity.getSupportActionBar();
+        if (actionBar != null) {
+            CharacterStyle span = new ForegroundColorSpan(activity.getResources().getColor(
+                    Themes.getResourceIdFromAttributeId(R.attr.actionBarTextColor)));
+            SpannableStringBuilder ssb = new SpannableStringBuilder(title);// Is it even necessary to use spannables anymore?
+            ssb.setSpan(span, 0, ssb.length(), 0);
+            actionBar.setTitle(ssb);
+        }
+    }
+
+
+    public static void setSubtitle(Activity activity, String title) {
+        setSubtitle(activity, title, false);
+    }
+
+
+    public static void setSubtitle(Activity activity, String title, boolean inverted) {
+        ActionBarActivity actionBarActivity = (ActionBarActivity) activity;
+        ActionBar actionBar = actionBarActivity.getSupportActionBar();
+        if (actionBar != null) {
+            if (inverted) {
+                CharacterStyle span = new ForegroundColorSpan(activity.getResources().getColor(
+                        Themes.getResourceIdFromAttributeId(R.attr.actionBarTextColor)));
+                SpannableStringBuilder ssb = new SpannableStringBuilder(title);
+                ssb.setSpan(span, 0, ssb.length(), 0);
+                actionBar.setSubtitle(ssb);
+            } else {
+                actionBar.setSubtitle(title);
+            }
+        }
+    }
 }
