+++ AnkiDroid/src/main/java/com/ichi2/compat/Compat.java
@@ -16,14 +16,10 @@
 
 package com.ichi2.compat;
 
-import android.app.Activity;
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
-import android.media.AudioManager;
 import android.speech.tts.TextToSpeech;
 import android.view.View;
-import android.view.ViewConfiguration;
-import android.webkit.WebView;
 import android.widget.RemoteViews;
 
 import com.ichi2.anki.exception.APIVersionException;
@@ -40,30 +36,21 @@ import com.ichi2.anki.exception.APIVersionException;
  * <p>
  * Each implementation should only override the methods that first become available in its own version, use @Override.
  * <p>
- * Methods not supported by its API will default to the empty implementations of CompatV7.  Methods first supported
+ * Methods not supported by its API will default to the empty implementations of CompatV8.  Methods first supported
  * by lower APIs will default to those implementations since we extended them.
  * <p>
- * Example: CompatV9 extends CompatV8. This means that the nfdNormalized function that uses classes only available
+ * Example: CompatV9 extends CompatV8. This means that the nfcNormalized function that uses classes only available
  * in API 9, should be implemented properly in CompatV9 with @Override annotation. On the other hand a method
- * like requestAudioFocus that first becomes available in API 8 need not be implemented again in CompatV9, unless the
- * behaviour is supposed to be different there.
+ * like setOverScrollModeNever that first becomes available in API 9 need not be implemented again in CompatV10,
+ * unless the behaviour is supposed to be different there.
  */
 public interface Compat {
     public abstract String nfcNormalized(String txt) throws APIVersionException;
     public abstract String detagged(String txt);
-    public abstract void setScrollbarFadingEnabled(WebView webview, boolean enable);
     public abstract void setOverScrollModeNever(View v);
-    public abstract void invalidateOptionsMenu(Activity activity);
-    public abstract void setActionBarBackground(Activity activity, int color);
-    public abstract void setTitle(Activity activity, String title, boolean inverted);
-    public abstract void setSubtitle(Activity activity, String title);
-    public abstract void setSubtitle(Activity activity, String title, boolean inverted);
     public abstract void setTtsOnUtteranceProgressListener(TextToSpeech tts);
     public abstract void disableDatabaseWriteAheadLogging(SQLiteDatabase db);
-    public abstract void requestAudioFocus(AudioManager audioManager);
-    public abstract void abandonAudioFocus(AudioManager audioManager);
     public abstract void enableCookiesForFileSchemePages();
-    public abstract int getScaledPagingTouchSlop(ViewConfiguration vc);
     public abstract void updateWidgetDimensions(Context context, RemoteViews updateViews, Class<?> cls);
 }
