+++ AnkiDroid/src/main/java/com/ichi2/compat/CompatV12.java
@@ -8,7 +8,7 @@ import android.webkit.CookieManager;
 public class CompatV12 extends CompatV9 implements Compat {
 
     // On API level 12 and higher, WebKit prevents file scheme pages from accessing cookies.
-	// This function removes this restriction.
+    // This function removes this restriction.
     @Override
     public void enableCookiesForFileSchemePages() {
         CookieManager.setAcceptFileSchemeCookies(true);
