+++ AnkiDroid/src/main/java/com/ichi2/libanki/Sound.java
@@ -92,6 +92,13 @@ public class Sound {
      */
     private static HashMap<Integer, ArrayList<String>> sSoundPaths = new HashMap<Integer, ArrayList<String>>();
 
+    /**
+     * Listener to handle audio focus. Currently blank because we're not respecting losing focus from other apps.
+     */
+    private static AudioManager.OnAudioFocusChangeListener afChangeListener = new AudioManager.OnAudioFocusChangeListener() {
+        public void onAudioFocusChange(int focusChange) {
+        }
+    };
 
     /* Prevent class from being instantiated */
     private Sound() {
@@ -307,7 +314,8 @@ public class Sound {
                     sMediaPlayer.setOnCompletionListener(playAllListener);
                 }
                 sMediaPlayer.prepareAsync();
-                CompatHelper.getCompat().requestAudioFocus(sAudioManager);
+                sAudioManager.requestAudioFocus(afChangeListener, AudioManager.STREAM_MUSIC,
+                        AudioManager.AUDIOFOCUS_GAIN_TRANSIENT_MAY_DUCK);
             } catch (Exception e) {
                 Timber.e(e, "playSounds - Error reproducing sound %s", soundPath);
                 releaseSound();
@@ -384,7 +392,7 @@ public class Sound {
             sMediaPlayer = null;
         }
         if (sAudioManager != null) {
-            CompatHelper.getCompat().abandonAudioFocus(sAudioManager);
+            sAudioManager.abandonAudioFocus(afChangeListener);
             sAudioManager = null;
         }
     }
